+++ platformio.ini
@@ -12,4 +12,4 @@
 platform = atmelavr
 board = uno
 framework = arduino
-monitor_speed = 9600
+monitor_speed = 115200
