+++ test/calibracionIMU.cpp
@@ -55,6 +55,9 @@ int16_t ax, ay, az, gx, gy, gz;
 int mean_ax, mean_ay, mean_az, mean_gx, mean_gy, mean_gz, state = 0;
 int ax_offset, ay_offset, az_offset, gx_offset, gy_offset, gz_offset;
 
+void meansensors();
+void calibration();
+
 ///////////////////////////////////   SETUP   ////////////////////////////////////
 void setup()
 {
